Added a "fake" wrapper for DataSorter that simply maps metric ID's between trials.  The mapping is done (once) by looking at the metric names as strings. This way, when two trials have the same 3 metrics, but they are in different orders such that metric 2 is PAPI_FP_INS in one, but is PAPI_L1_DCM in the other, they will correctly map to the right named metric.  It's a bit of a kludge and should replaced with a better method at some point.
There is still a problem if the first trial has a metric than one of the others
does not.  I need to find a way to have this map to a set of all 0s.  Right
now, if it can't find the metric on the other trial, it will just go with
metric 0.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4376 a8e31b63-e6ca-409b-a01b-2eec1b506579