This fix solves two bugs:
PDT 0000163:
Write a identicalBeginEnd() routine that returns true if a routine's body begin has the same location as its body end. This is noticed when a macro expands to a routine name as in Uintah:

DECLARE_MAKER(CreateMatrix)
CreateMatrix::CreateMatrix(GuiContext* ctx)
  : Module("CreateMatrix", ctx, Source, "Math", "SCIRun"),
    nrows_(get_ctx()->subVar("rows"), 1),
    ncols_(get_ctx()->subVar("cols"), 1),
    data_(get_ctx()->subVar("data"), "{0.0}")
{
}

It gets instrumented as:
DECLARE_MAKER(CreateMatrix)
CreateMatrix::CreateMatrix(GuiContext* ctx)
  : Module("CreateMatrix", ctx, Source, "Math", "SCIRun"),
    nrows_(get_ctx()->subVar("rows"), 1),
    ncols_(get_ctx()->subVar("cols"), 1),
    data_(get_ctx()->subVar("data"), "{
  TAU_PROFILE("SCIRun::Module *SCIRun::make_CreateMatrix(SCIRun::GuiContext *) C [{CreateMatrix.cc} {36,1}-{36,1}]", " ", TAU_USER);
0.0}")
{
  TAU_PROFILE("void SCIRun::CreateMatrix::CreateMatrix(SCIRun::GuiContext *) [{CreateMatrix.cc} {37,1}-{43,1}]", " ", TAU_USER);

}

and
PDT 0000162: tau_instrumentor: main(char **, int ) not instrumented correctly
Description	 TAU_INIT(&-, &-); is called when main does not have argc, argv. We should not invoke TAU_INIT if this is the case.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4617 a8e31b63-e6ca-409b-a01b-2eec1b506579