Phase 1 of PERI/XML_Metadata/whatever support: perfdmf_loadtrial now accepts another parameter:
-m filename
--metadata filename

The parser will check the database to see if the XML_METADATA column
exists in the Trial table.  If not, it will create it:
derby, oracle, db2: CLOB
mysql, posgresql: TEXT

Because it is stored as character data (and not binary), I don't have
to do any craziness to read it back in.  I did NOT make it a binary
type, because in JDBC, each of the 3 databases I tested with (mysql5,
postgresql and derby) had different binary types, and all handled the
data inconsistently.  ASCII data is all handled correctly.

Once the code confirms that the column exists in the database, the code
will parse the XML file using statement.setAsciiStream, which takes a
FileInputStream and File.length() as parameters.

When the data is read and displayed in paraprof and perfexplorer, it
looks like a long string of XML data.  The next phase is to put that
data into the SAX parser, and parse it.

Potential todo for this code: validate the XML metadata when it is
first put in the database.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4634 a8e31b63-e6ca-409b-a01b-2eec1b506579