Currently, we can't handle statements like: if (value .gt. 3) &   allocate(A(2))
So, we put in an error message:
TAU ERROR: <file=bar8.f90,line=7>: Currently we cannot handle allocate statements that are on the same line as a single-if statement that uses a continuation character. Please modify the source to put an explicit then/endif clause around the allocate statement and re-try.
TAU ERROR: <file=bar8.f90,line=9>: Currently we cannot handle de-allocate statements that are on the same line as a single-if statement that uses a continuation character. Please modify the source to put an explicit then/endif clause around the de-allocate statement and re-try.
       subroutine bar(value)
       integer:: value, err
       integer, dimension (:),  allocatable :: A


      integer profiler(2) / 0, 0 /
      save profiler

      call TAU_PROFILE_TIMER(profiler, 'BAR [{bar8.f90} {1,19}]')
       call TAU_PROFILE_START(profiler)

       allocate(A(2))
       call TAU_ALLOC(A, 5, sizeof(A), 'bar8.f90, variable=A')

       if ( value .gt. 3) &
         allocate(A(3))
       if ( value .gt. 3) &
         deallocate(A)

       call TAU_PROFILE_STOP(profiler)
       end subroutine bar


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4849 a8e31b63-e6ca-409b-a01b-2eec1b506579