We can now handle:       subroutine foo(x)       integer:: x
      integer, allocatable:: A(:), B(:)

      if (x .gt. 2) &
        allocate(A(2))

      deallocate(A)

      allocate (A(2),  &
       B(2))

      if (x .gt. 2) &
        deallocate(B, &
          A)

      deallocate(A, &
        B)
      end subroutine foo

	There was a problem with the last statement of the routine. A copy
of the last line was being maintained (in previousline). This was not correctly
updated when the allocate/deallocate print routines read additional lines. So,
the return statement thought it was part of an if/then clause and treated it as
a
	if (x .gt. 3) &
	  return
type of a statement and added additional then/endif statements. Keeping a copy
in previousline and updating inbuf accordingly resolves this problem.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4932 a8e31b63-e6ca-409b-a01b-2eec1b506579