Added runtime clock synchronization for TAU traces.  Use -useropt=-DTAU_SYNCHRONIZE_CLOCKS to use.  We can now use LINUXTIMERS for the time source in traces, and NTP is no longer necessary on a cluster.
Synchronization takes place during MPI_Init() with ping/pong messages.
Additionally, clocks are reset to 0 with the caveat that all records before
MPI_Init() will have time 0.  We should probably change this in the future to
modify the records in the buffer (up to 64k records) with time offset
adjustments instead.

The idea and some code is adapted from Kojak.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4984 a8e31b63-e6ca-409b-a01b-2eec1b506579