Fixed insidious casting bug in the Tracer.  It was possibly to produce a trace where time went backward with clock synchronization.  Some parts of the code were casting the double from getUsecD to an x_uint64 before calling the adjust routine.  C appears to truncate this:
Message A       : time = 4.7
end of MPI_Recv : time = 4.8

If the Stop of MPI_recv casts it to x_uint64 before adjusting it, it is reduced
to 4.0 and can end up before "Message A".

This bug has been exposed through the low overhead linux timers.  I don't think
it would be possible with getttimeofday, since two successive calls will be
several microseconds apart.

The real bitch comes when tau_merge goes to merge traces with an event like
this.  If tau_merge ever sees a timestamp that is less than the previous, it
will assume that the clock has overflowed and the rest of that trace will come
after all the others.  Vampir and other tools will then barf all over the place
since the merged TAU trace has time going forward and backward all over the
place.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4988 a8e31b63-e6ca-409b-a01b-2eec1b506579