We now check if the jvmpi interface has been initialized when the Env mutex is created.  If it has not, TAU will output:
Error, TAU's jvmpi interface was not initialized properly (java -XrunTAU ...)
When TAU is configured with -jdk=<dir>, it can only profile Java Programs!

and then exit(-1), since the next line will segfault (null pointer).

The usual cause is that the user has specified -jdk=<dir> and did not know that
this specifies a Java-only measurement library.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@5173 a8e31b63-e6ca-409b-a01b-2eec1b506579