Added TAU_ALT_FORTRAN_INSTRUMENTATION for nagware and g95.
We may want to automatically turn this on when these compilers are used.  But,
you can only have one tau_instrumentor binary even if you build TAU with many
compilers.  So I think we should end up going with Sameer's suggestion and make
it a flag to tau_instrumentor which is automatically set with stub makefiles
for nagware and g95.

Alternatively, we should try this alt instrumentation on all the other fortran
compilers (including many version of each).  If they all take it, then we just
swap it in for the regular and we're done.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@5410 a8e31b63-e6ca-409b-a01b-2eec1b506579