Instead of hardcoding the full path to the original header, we can use: #include <../header.h>
for a given header.h in a wrapper redirecting header.h. Then, it automatically
includes the original header and we can get all the types etc. that were defined there. The types are needed for making a prototype for
extern <rettype> tau_foo(args...);
for a given foo routine.

So, we will need to maintain a copy of the original header within the component directory (and the wrapped header lies in the wrapper/ subdirectory).


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@5587 a8e31b63-e6ca-409b-a01b-2eec1b506579