Added smart clustering to the PerfExplorer script interface.
What's smart clustering you ask?

see: "Estimating the number of data clusters via the Gap statistic", by Hastie,
Tibshirani and Walther, 2000. (or 2001, depending on the version of the paper)

First, try k-means clustering with 1 cluster.
	Compute the "Gap Statistic" for 1 cluster.
Then, try k-means clustering for 2 clusters.
	Compute the "Gap Statistic" for 2 clusters.
	If GS2-error > GS1, continue.
Then, try k-means clustering for 3 clusters.
	Compute the "Gap Statistic" for 3 clusters.
	If GS3-error > GS2, continue.
...and so on.

Break when the new GS is not greater than the previous GS, and return the
cluster results from the previous run.

I have tried it with some data sets I have, and it sort of works.  I am
not confident the GS is computed accurately, yet.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@6414 a8e31b63-e6ca-409b-a01b-2eec1b506579