Fixed instrumentation of main() for C++ in '-spec' mode:
In C++, the return statement at the end of main() is optional (although
it has return type int). However, only return statements were instrumented
for non-void functions if exit instrumentation was requested. This has now
been changed to also instrument the end of main() as a special case. This
leads to some dead code if a return statement is present, but who cares...
;-)


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@6602 a8e31b63-e6ca-409b-a01b-2eec1b506579