Added 'abort' construct:
    abort  file="..."  routine="..."  code="..."  lang="..."

Instruments all locations in the source code where
  - 'exit()' or 'abort()' is called (C/C++)
  - 'stop' is used (Fortran)
This construct should work in all three modes: TAU, PerfLib and '-spec'.
For C/C++, currently only real calls to exit()/abort() are recognized,
i.e., alternative keywords provided on the instrumentor's command line
are not yet supported.
Keyword substitutions in the code snippet are the same as for 'entry' etc.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@6610 a8e31b63-e6ca-409b-a01b-2eec1b506579