Patched with Davide Del Vento's patch. (<ddvento@ucar.edu>)
Date: Wed, 05 Aug 2009 16:51:26 -0600
From: Davide Del Vento <ddvento@ucar.edu>
To: tau-bugs@cs.uoregon.edu
Subject: tau_validate
Parts/Attachments:
   1 Shown     83 lines  Text
   2 Shown    291 lines  Text
   3 Shown    241 lines  Text
----------------------------------------

Dear Sirs,
first, I'd like to thank you for your efforts in writing and making TAU
available to the community.
Then, I'd like to report a behavior of the tau_validate script, which I
consider a bug - I've also included a patch for it.
Last, but not least, I'd like to submit a separate patch for the same
tau_validate script, which in my opinion, would make its use more
convenient.

***** But let's start with the bug *****
In line 82, within the system() function, there is a call to the wait()
method. This means that if the program under test hangs, for whatever
reason, the whole tau_validate suite hangs as well. This actually
happens in our current TAU installation, which probably contains
something broken. I think that the tau_validate script should complete
always, even in case of a broken installation, reporting which tests
hanged (I left some ./simple programs running for a hour, before
deciding that it was hanging - our machine is bluefire, an IBM Power6
running at 4.7GHz).
To solve this problem, I designed a timeout, implemented with the signal
module. See the attached patch.timeout
The core of the changes are in the system() function. A crucial but a
little cryptic issue is the command = command.split() statement
performed in case of timeout. This is due to a odd but documented
feature of popen2.Popen4, described here:
http://docs.python.org/library/popen2.html

> On Unix, cmd may be a sequence, in which case arguments
> will be passed directly to the program without shell
> intervention (as with os.spawnv()). If cmd is a string
> it will be passed to the shell (as with os.system()).

In fact, if I wouldn't split the string, it would run in a shell, and
the os.kill() call in case of timeout, would kill the shell, which on
many unix's would NOT cause the ./simple program to be terminated: the
tau_validate python program would correctly running, but the test
executable would not stop. Before realizing it, I quickly flooded our
system!

The rest of the patch is just boilerplate to accept a timeout value from
commandline and propagate it to the actual method. And of course a small
change to the outputSummary() method, to report in HTML the timeout
results also, besides failures, successes, and n/a's. Note that I'm not
using the timeout for operating system activities or for the build
process which should never hangs. In fact, using the timeout, might
cause problems to these, since the missing shell could mean other
missing things (for examples the pipes "|" would not work).

***** The second patch is for the way the output is produced *****

A missing feature of the tau_validate program, is that it produces
the very useful table only at the end of the run.
It is possible to monitor the course of the run while the suite is
running, but only from the log, which (in my opinion) is very
inconvenient. Thus I wrote additional code (you can read it in the
attached patch.table file) which creates a table-file, identical to the
summary usually printed at the end of the run. This table-file, is
created "on-the-go", as soon as the result of each test is ready.
The core of the patch is trivially the split of the outputSummary()
method in a outputSingle(). The latter method is  called by the
outputSummary() itself, but also at the end of the main program loop, in
line 812. I could have split the output on a even more fine-grained
fashion (e.g. single individual tests), but I found waiting for a whole
TAU Makefile not too long, and simpler to implement.
To avoid cumbersome syntax in the print statements, the trick is to
automagically redirect the stdout (see line 811) to the file created in
line 806. The rest, again, is boilerplate for command line argument parsing.
Note that this patch include also some minor HTML fixes, such as
lowercase tags (see here f


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7337 a8e31b63-e6ca-409b-a01b-2eec1b506579