Fixed critical bug in parameter profiling causing the map of parameters to be nearly useless and making this basically a dynamic timer that is re-created each time.
The code creates arrays of longs with:

long * TauCreateProfileParamArray(long FuncId, long key) {
  int depth = 3;
  long *retary = new long[depth+1];
  retary[0] = depth; /* encode the depth first */
  retary[1] = FuncId; /* the id of the current timer */
  retary[2] = key;   /* data */
  return retary;
}


And the comparison operater is defined as:

struct TaultProfileParamLong {
  bool operator() (const long *l1, const long *l2) const {
   int i;
   /* first check 0th index (size) */
   if (l1[0] != l2[0]) return (l1[0] < l2[0]);
   /* they're equal, see the size and iterate */
   for (i = 0; i < l1[0] ; i++)
   {
     if (l1[i] != l2[i]) return l1[i] < l2[i];
   }
   return (l1[i] < l2[i]);
 }
};


The bug here is that the depth is hardcoded incorrectly.  If the two arrays
have the same depth, FuncID and key, they may or may not be identified as equal
since the comparison function iterates from 0 to 3, and i will always be 4 for the call:

   return (l1[i] < l2[i]);

l1[4] and l2[4] are valid memory addresses, but are always uninitialized.

To fix this, we simply change the depth to '2'.  Additionally, the comparison
operator is checking l1[0] and l2[0] twice for some reason, this has been fixed
as well.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7353 a8e31b63-e6ca-409b-a01b-2eec1b506579