Checking in changes to EBS necessary to:
* Support Linux on PPC970 (MareNostrum) - but it crashes after the first sample
* Support deltas for hardware counters
* TAU_EBS_FREQUENCY environment variable for changing the number of
  microseconds between each sample
* Post-process deltas for hardware counters
* Extract samples from the post-processed sample files as a first step
  towards folding iterations

the format could still use some optimization, as in the final post-processed
samples, the first metric in the list of metrics has deltas for TIME,
which is the same as the deltas for the timestamp of the sample.

Alan, I am also hoping that if you have some spare cycles (yeah, right),
you might be able to help debug the sampling on MareNostrum.

We are still debugging the samples - we get crazy outliers every
once in a while.  Also, there are some "philosophical differences" with
regard to which function a sample should belong to, when it comes to
folding iterations for those functions.  For example, when
folding, all samples should belong to application code functions, not
any possible children.  In the BT benchmark, when COPY_FACES makes a call
to MPI_Wait, and a sample is taken, the sample should belong to COPY_FACES,
not MPI_Wait.  Otherwise, there are gaps in the samples for COPY_FACES
where the samples are "owned" by MPI_Wait.
This can't be "fixed" in the post-processing, the sampler has to know that
the counter values that the user wants are for the parent function,
not the MPI call.

Hope that makes sense.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7527 a8e31b63-e6ca-409b-a01b-2eec1b506579