Some bug fixes, lots of new code.
First, fixed a bug in Hierarchical clustering, I doubt it ever worked,
although no one noticed.  The distance matrix weights are in a separate
vector now, not in the main matrix, and the main matrix has redundant data,
but the processing is simplified.

Second, added new operations: SplitCommunicationComputationOperation (does
what it says), and HierarchicalClusterOperation (to access it from scripting).

When this is all finished, we should be able to request hierarchical
clustering directly from the GUI, too.

I am still debugging the kmeans clustering and hierarchical clustering,
because I am getting different results from the GUI and scripting, which
makes no sense because they (theoretically) use the same code.  I wanted
to check in my work for now, so I could access it from home.  there are some
debugging messages in there which should be cleaned out later, so if you
see some noise from clustering, that's why.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7534 a8e31b63-e6ca-409b-a01b-2eec1b506579