Checking in latest changes for sampling and folding.  Hopefully this is it. I added an environment variable, TAU_EBS_INCLUSIVE, which will output a sample for all functions in the callstack when a sample is taken.  In order to post-process these samples properly, call the tausamp_process.pl script with either "-i" or "--inclusive".  Also, the tausamp_extract4folding.py script has been updated to reflect the latest changes with Harald's folding code.
I don't know whether the folding code will be included in the TAU CVS
(probably not), but when this phase is finished, I will move on to finishing
the code in paraprof to launch gnuplot when this data exists.

Alternatively, we could try to implement the "gnuplot" part with jfreechart,
but the folding generates gnuplots with 1) a scatterplot 2) a projected
curve 3) a projected curve which represents the slope of the first projected
curve.  Not sure if Jfreechart can handle all that.  We shall see.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7564 a8e31b63-e6ca-409b-a01b-2eec1b506579