PDT has a really hard time parsing C files on Macs. C++ files are no problem. So, we had a a edgcpfe --g++ argument just for Intel Macs. This messes up tau_instrumentor for C files. When we use: % cparse foo.c % ls foo.pdb it generates foo.pdb properly, but then it has a lang c++
instead of
lang c
as the entry. Now, tau_instrumentor reads the lang c++ and instruments it with a C++ TAU_PROFILE macro in the foo.inst.c file. Since there are no constructors or destructors in C, on Intel Macs (where we had a --g++ argument), we get no instrumentation in a .inst.c file as TAU_PROFILE is defined as
null for C. So, to get around this problem, we inject a -optTau=-c parameter for C files in tau_cc.sh. tau_cxx.sh is unaffected by this change. Now -c appears in the tau_instrumentor commandline and it can instrument the C++ pdb file and C file as a C file while writing out foo.inst.c.  So, all the C entities now appear properly in the profiles. Weird! Since -optTau=-c does not reset the options but rather appends to them, if the user defines TAU_OPTIONS '-optTau="-i TAU.h" -optVerbose' for instance, tau_instrumentor gets both -c and -i TAU.h on the commandline. So, everything works properly now on Snow Leopard for both C and C++.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7622 a8e31b63-e6ca-409b-a01b-2eec1b506579