Now that Sun owns MySQL, they've changed the license.  According to this page:
http://www.mysql.com/about/legal/licensing/foss-exception/

We can now distribute the JDBC driver.  They allow for an exception for BSD and
other "compatible" licenses.

I also checked the PostgreSQL website and their JDBC driver is now BSD licensed.

So now we can distribute both of these jars and don't need to do all the
downloading junk!


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7667 a8e31b63-e6ca-409b-a01b-2eec1b506579