Fixed bug with percentages and mean values:
Howdy -

Attached is a profile which demonstrates a bug in ParaProf.

When showing the mean value of do_nonbonded (exclusive TIME) as a percentage, the value should be either 23.91% or 16.67%, depending on the value of the preference that controls how to deal with processes with 0 calls to that function (do_nonbonded is not called by all processes).

However, regardless of how the preference is set, the percentage is always 16.67%.  The *value* however, does change, from 97.99 seconds to 68.273 seconds.  The inclusive time for main is always the same, so this seems to be a bug, right?

Thanks -
Kevin


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7797 a8e31b63-e6ca-409b-a01b-2eec1b506579