Fixed problem reading "counting" psrun data files:
Attached is a PerfSuite HWPC counting report that paraprof in TAU 2.19 has problem loading, while paraprof in TAU 2.18.1 does not. The difference between it and your test file is that this file has 2 more elements: <pid> and <thread>, between the elements <domainname> and <username>. The <pid> and <thread> elements are defined as optional in the hwpcreport.dtd file:

<!ELEMENT hwpcreport ( date,
                       hostname,
                       domainname,
                       pid?,
                       thread?,
                       username?,
                       command?,
                       annotation?,
                       wallclock,
                       cputime,
                       machineinfo?,
                      (hwpceventdata | hwpcprofiledata)
                     )
>

Therefore, your test file is valid; but the attached file, which was generated file by running an application, is also valid. So it might be desirable to be able to support both.

Please let us know if you have more questions. Thanks!


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@7846 a8e31b63-e6ca-409b-a01b-2eec1b506579