Initial version of ToM. Has built-in (non-modular) protocols to:
1. compute basic statistics
2. compute histograms

Activated by configure options (Both options required). If options
not enabled, regular TAU operations should function normally:

-mrnet=<mrnet root dir>
-mrnetlib=<mrnet lib dir>

This is my first major check-in, please be patient and let me know if
it screws anything up. This has been tested on mist.cs.uoregon.edu with
a simple hello code for ToM operations and NPB for regular TAU operations.
TauCollate continues to work correctly in response to TAU_ONLINE_DUMP().
I plan to test a clean checkout of this code on kraken as soon as I can.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@8278 a8e31b63-e6ca-409b-a01b-2eec1b506579