Using new technique for dealing with calloc.  The primary issue is that when wrapping calloc, we need to use dlsym to get to the real calloc, but dlsym can sometime use calloc itself!  So a niave approach results in an infinite loop.
Our previous attempt to address this involved returning a pointer to a static
space for the first calloc call.  But this approach has several drawbacks.  For
one, we can't really predict where the first calloc call will come from.  We
tried to address that by forcing a calloc call in the base tau preload library,
and always enable it with tau_exec.  However, as the dyninst example
demonstrates, many calloc calls can occur before our dlsym call to calloc is
ever made, so we were stuck once again.

The new method involves wrapping calloc in such a way that the first invocation
will set a flag to be checked by subsequent calls, then forcing the dlsym call.
This dlsym call may or may not call calloc, but if it does, then the wrapper
will see the flag and invoke the previous technique of returning the address of
the static space.  If it doesn't invoke calloc, then the static space will
never be used.

This fixes the dyninst case:

cat f1.f90
      program main
      include "mpif.h"
      integer ierr

      call MPI_INIT( ierr )
      call MPI_FINALIZE(ierr)
      end program main

mpif90 f1.f90 -o f1

tau_run f1 -o f1.i

tau_exec -memory ./f1.i


And also passes 100% of the wrapsuite test cases.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@8300 a8e31b63-e6ca-409b-a01b-2eec1b506579