Initial commit of JVMTI. Needs to be integrated into the build system.
TauJVMTI.[cpp|h] Interfaces witht he interpreter using JVMTI and performs the rewriting of the class files to produce the injected bytecode.
agent_util.[ch] Contains helper code for managing the JVMTI interface.
java_crw_demo.[ch] Is a great library for rewriting bytecode
TauJVMTI.java is a java file that calls back methods in TAUJVMTI.cpp since the injected method calls must be to actual Java methods, not native ones this performs the necessary indirection.

There is some legacy code in the TauJVMTI files that prints non-Tau integrated profiling data to the screen that should be removed once we have proven it works the same as TAU.
Threading support needs to be improved.
Provisions for tracking the memory and IO can be added