Added TAU_USE_GPU tag for slightly cleaner dependency handling in our makefile system.
Temporarily fixed a dependency problem between -mpi and -cuda. On a clean
system, when both are used together on the x86_64, -cuda prevents the
necessary MPI library from getting installed into the appropriate lib
directory. This does not happen on the apple build. Further investigation
into dependency consistency is required. The "fix" involved the swapping
of the positions of CUDA and MPI installation dependencies in the
makefile. This implies that "make -j install" may still fail.