A working version of BFD name resolution for Comp_gnu with a shared code base used by TauSampling as well. Function names in shared objects should be correctly resolved on the platforms listed below as tested.
This has been manually tested on the
Cray XE6 (hopper), BGP (intrepid), x86_64 (mist) and i386 (desktop) over
serial and MPI codes and various build (compiler etc) environments. This
has NOT been tested with threads. I'll be relying on the regression to
systematically expose the inevitable bugs as they show up.

This check-in also provides preliminary functionality for Profile-based
Sampling with online name resolution using BFD. Name-resolution functionality
is similar to that supported for Comp_gnu, with the exception of having extra
features to handle arbitrary sampled addresses. Profile-based Sampling is
always built. Default behavior is not to sample. Turning it on only requires
that the TAU_SAMPLING environment variable be set to 1.

This commit also includes various bug fixes in the way TAU is built for the
BG/P.