Added a test in the configure script to determine if the compiler used "likes" the funky C++ code in the special storage allocator used for the purposes of sampling. If it does not, then the default STD allocator is used instead.
The code also paves the way for additional flags to handle compilers
that are ok building the code, but executes them in unexpectedly bad
ways. These flags would include those for PGI and AIX compilers. This
would allow sampling to for most cases until we can find proper fixes
in time.


Former-commit-id: 43beb44a660dc673c76c397f512119e9b28d69de