Added the necessary event entries to allow Sampling-generated constructs to work correctly when used with Reverse Call Tree ParaProf functionality.
Also added stub checks for a request for Tau Merged format in MPI_Finalize.
In the context of TAU Sampling, we would want to resolve the symbol
information and create the FunctionInfo objects before the call for
unification so the merged output will actually contain sampled
information.

There are current issues with threading that do not permit an easy
way for this to happen as yet.