Cleaned up the way EBS handles TAU_EBS_SOURCE other than "itimer". Integrated EBS Profiles will now correctly write approximated sampled metric values (based on the EBS_SOURCE) to the appropriate metric column instead of being restricted to just TIME. This is done through a new interface for getting the Metric vector index using the metric name inside TauMetrics.cpp.
Fixed a bug where we fail to prevent the sampling handler from being invoked
by PAPI overflow interrupts when we are done (via samplingEnabled flag).
The full solution is to turn off PAPI interrupts when that happens in addition
to setting the flag, but it is unclear if this is a simple thing to do.

As of this check-in, neither timer interrupt nor PAPI overflow signals are
being correctly delivered to threads other than 0.