Added auto BFD path detection for Cray CCE compiler on the craycnl architecture. If BFD is to be implicitly discovered (in other words, no "-bfd=<path>" is specified), a failure to locate CCE's binutils path (as defined in the environment variable CRAY_BINUTILS_ROOT when the cce module is loaded) will result in a warning message that implicit BFD support has failed instead of passing any discovered BFD link flags to the TAU build.
The reason this rather extreme decision is taken is because the CCE
compilers, as they are currently set up on craycnl, defines different
library search paths for when the -shared flag is used. As a result,
when the configure script is run, the -lbfd etc ... flags are correctly
(and implicitly) discovered BUT will fail to be found at TAU build time.
This results in TAU's failure to build.


Former-commit-id: 2b105ee7a427bc2178c2e68ef88fc8f0315c397c