3 changes/bugfixes for/because of Mac OS X:
1. The ability to link against librt needs to be checked before attempting
to use -lrt in TAU_LINKER_OPT16. To this end, a new check has been added to
update the TAU_LINKS_RT tag in the configure script. This has also prompted
two refinements - the first is to avoid wasting time checking for RT Clock
Signal functionality for EBS if librt wouldn't even link. The second was to
remove the unnecessary attempt to restore EBS_HAS_RT and TAU_SS_ALLOC_SUPPORT
values in FixMakefile.sed.default. Those two symbols are tags and not
variables.

2. No RT interface support exists for Mac OS X (it is NOT POSIX compliant!),
calls made to clock_gettime in TauReadMetrics.cpp have to revert to using
gettimeofday. Including time.h does not help on the Mac unlike the case of
the PGI compiler on mist.

3. The check for __APPLE__ in iowrap_link.c has been moved to the top so
it is now much safer to add any new functionality that requires the
"off64_t" type.

Fixed a long-time build/git problem with Mac OS X involving APPLECXX and
the EXTRA_LINKER_ARG variable. Please read include/Makefile for details.
As an aside, we still have two outstanding configure/git issues in the
form of Mac OS X's non-case-sensitive approach to file names and as a
result, its refusal to play nice with the names "INSTALL" and
"include/Memory". Right now, we hack configure to rename "INSTALL" and
delete "include/Memory" but fail to restore it upon the invocation of
configure default. I currently do not know of a good way of dealing with
this that does not involve hardcodes or deeper hacks to the build system.
Developers of TAU using Mac OS X as their main development platform will
have to be very very careful with checkins for now. Please do not
perform blanket checkins with git if you are developing from a Mac OS X
system.