Multiple fixes here:
1. Added Rui's fix for intel 12 to configure. Here are the details:

We are preparing a new machine named "Forge" at NCSA for a
friendly-user mode.  I was installing and testing TAU 2.20.3 on it.
Testing of the "f90" example failed in linking. Debugging found that
it used an incorrect path for the "for_main.o" file, and the root
cause was the logic in the "configure" file for setting the
IFORTLIBDIR variable.

For Intel compiler 11.x, the path to ifort is:
   /usr/local/intel/Compiler/11.1/073/bin/intel64/ifort
and the "for_main.o" is in:
   /usr/local/intel/Compiler/11.1/073/lib/intel64/.

The existing logic in "configure" worked well for this case.

However, for Intel compiler 12.x on Forge (we have 12.0.4), the
directory structure changed.  The path to ifort is:

   /usr/local/intel/bin/ifort
while the "for_main.o" is in:
   /usr/local/intel/lib/intel64/

The existing "configure" logic caused the IFORTLIBDIR to be set to
/usr/local/intel/lib, and this led to a link time failure.

2. Switched from an arbitrary flag-based determination of whether we
could/should make an extern declaration for the non-portable strsignal
function to one that is determined by the configure script.

Personal note: We need to try to minimize the use of non-portable
functionality. If we do, we ought to make the detection of non-portable
support a part of the configuration script - that way, the pain of
a slow configuration will remind us of when we need to review the use
of such features. Arbitrary use of conditional flags just gunk up the
code and throw developers off, particularly since they tend to be
poorly documented and may change over time.