Some fixes on the Cray XE6 system (hopper):
1. CCE compilers cannot appear to handle -lmpichcxx anymore for shared builds.
They require -lmpichcxx_cray.

2. Built in a temporary workaround for Intel compilers on hopper - added an
option to override BFD by turning it off for pathological cases like these.
In this case, libiberty.h is incompatible with stdio.h for the PrgEnv-intel
module on hopper. This issue has been raised with consult@nersc and they
have forwarded the information to Cray engineers.

3. This is not technically part of the checkin, but PrgEnv-pathscale on
hopper is also broken - include <string> will fail and TAU uses that. As
such, TAU will not build under this module on hopper (it has, previously).
The problem has been noted to consult@nersc. I will be waiting for a reply.