Changes to the way EBS uses templated code. The Pathscale compilers on hopper are very particular about how STL maps are used. Essentially, the specification of pair has to be of type <const key, value> rather than the less strict use of <key, value>. I agree with this change and this should put us in good stead with other strict compilers.
However, Pathscale is also messed up where it comes to handling the
optional comparator and allocator fields of map. It is also broken
when handling key = char * (this is not a direct problem of our code,
but is certainly something to pay attention to).

In any event, this version of the code will allow TAU to be built
successfully for all five major compilers on hopper (pgi, gnu, pathscale,
cray and intel). EBS will successfully run and generate correct output
for pgi, gnu and cray. Something else (not EBS) is preventing proper
testing of EBS on the NPB benchmarks for the pathscale and intel
compilers and will require additional debugging.


Former-commit-id: 68c7b4781dc326d1465ee4d7cf810f655004e89c