Extra safety checks in Comp_gnu.cpp.  Removed dead code from TauBfd.cpp.
GNU compiler-based instrumentation callback function was reading
TauBfdInfo fields without first checking to see if they were non-NULL.
This is correct because calling Tau_bfd_resolveBfdInfo() on a TauBfdInfo
struct will set all fields to be non-NULL, even if the info is not
fully resolved.  Even so, this update adds code to verify that the fields
are non-NULL before reading in case something changes in future.