When rose is used to parse the source code, it does not generate a PDB record for the last return statement in a void function. This leads to timer overlaps because the timer for the void function is not stopped right before this last return. The bug is in EDG 3.3. It does not affect PDT's cxxparse because it uses a different version of EDG software. How do we fix this return bug?
We check if -optUseReturnFix is used or if roseparse or upcparse is passed
to the script. tau_upc.sh does this. Then, we change the source code:
mv return1.c return1.c~; sed -e 's@return\([         ]*\);@{return \1;}@g' return1.c~ > return1.c;

and after instrumentation, we use:
mv return1.c~ return1.c

So, BEFORE:
cat return1.c
 int returnsInt() {
  int a = 1;
  return 2;
 }

 void returnsVoid() {
  int a = 1;
   if (a >2) return      ;
   if (a >3) return              ;
  a=3;
  return ;
 }

and AFTER:
 int returnsInt() {
  int a = 1;
  return 2;
 }

 void returnsVoid() {
  int a = 1;
   if (a >2) {return       ;}
   if (a >3) {return               ;}
  a=3;
  {return  ;}
 }

By putting return in the { } block, rose generates the correct pdb records for this statement and the instrumented file has:
 void returnsVoid() {

        TAU_PROFILE_TIMER(tautimer, "void returnsVoid()  C [{return1.c} {7,2}-{13,2}]", " ", TAU_USER);
        TAU_PROFILE_START(tautimer);

{
  int a = 1;
   if (a >2) {{  TAU_PROFILE_STOP(tautimer);return        ; }
}
   if (a >3) {{  TAU_PROFILE_STOP(tautimer);return                ; }
}
  a=3;
  {{  TAU_PROFILE_STOP(tautimer);return   ; }
}

}

        TAU_PROFILE_STOP(tautimer);

}

Now, when cparse internally falls back to using the roseparse parser, there is no way to detect this internally in TAU (short of parsing the pdb file to check - but by then it is too late). So, we provide the -optUseReturnFix.


Former-commit-id: 5c7d4c6a9d9550fb70d969c422ec6202db66727d